:recycle: Move hwWallet HOC to seperate file
